Add kernels.bonds.py and refactor
