+++ pypd/particles.py
@@ -8,8 +8,12 @@ TODO: rename classes as base or baseclasses?
 import numpy as np
 import matplotlib.pyplot as plt
 
-from .tools import build_particle_families, smooth_step_data
-from .kernels.particles import calculate_nodal_forces, calculate_node_damage
+from .tools import smooth_step_data
+from .kernels.particles import (
+    build_particle_families,
+    calculate_nodal_forces,
+    calculate_node_damage,
+)
 
 
 # Particles, ParticleArray, or ParticleSet?
