+++ pypd/tools.py
@@ -1,8 +1,5 @@
 import numpy as np
-import sklearn.neighbors as neighbors
-from numba import njit, prange
-
-# TODO: should this be a class (Particles)?
+from numba import njit
 
 
 @njit
@@ -18,111 +15,6 @@ def smooth_step_data(
     )
 
     return alpha
-
-
-def build_particle_coordinates(dx, n_div_x, n_div_y, n_div_z):
-    particle_coordinates = np.zeros([n_div_x * n_div_y * n_div_z, 3])
-    counter = 0
-
-    for i_z in range(n_div_z):  # Width
-        for i_y in range(n_div_y):  # Depth
-            for i_x in range(n_div_x):  # Length
-                coord_x = dx * i_x
-                coord_y = dx * i_y
-                coord_z = dx * i_z
-                particle_coordinates[counter, 0] = coord_x
-                particle_coordinates[counter, 1] = coord_y
-                particle_coordinates[counter, 2] = coord_z
-                counter += 1
-
-    return particle_coordinates
-
-
-def build_particle_families(x, horizon):
-    """
-    Build particle families
-
-    Parameters
-    ----------
-    x : ndarray (float)
-        Material point coordinates in the reference configuration
-
-    horizon : float
-        Material point horizon (non-local length scale)
-
-    Returns
-    -------
-    nlist : list of numpy arrays (int)
-        TODO: define a new name and description
-
-    Notes
-    -----
-    TODO: include a discussion of the algorithm
-
-    """
-    n_nodes = np.shape(x)[0]
-
-    tree = neighbors.KDTree(x, leaf_size=160)
-    neighbour_list = tree.query_radius(x, r=horizon)
-
-    # Remove identity values, as there is no bond between a node and itself
-    neighbour_list = [neighbour_list[i][neighbour_list[i] != i] for i in range(n_nodes)]
-
-    n_family_members = [len(neighbour_list[i]) for i in range(n_nodes)]
-    n_family_members = np.array(n_family_members, dtype=np.intc)
-
-    nlist = np.ones((n_nodes, n_family_members.max()), dtype=np.intc) * -1
-
-    for i in range(n_nodes):
-        nlist[i, : n_family_members[i]] = neighbour_list[i]
-
-    nlist = nlist.astype(np.intc)
-
-    return nlist, n_family_members
-
-
-def build_bond_list(nlist):
-    """
-    Build bond list
-    """
-    bondlist = [
-        [i, j] for i, neighbours in enumerate(nlist) for j in neighbours if i < j
-    ]
-    bondlist = np.array(bondlist, dtype=np.intc)
-
-    return bondlist
-
-
-def build_penetrator():
-    """Build rigid penetrator"""
-    pass
-
-
-def build_volume_correction_factors():
-    pass
-
-
-@njit(parallel=True)
-def build_bond_length(x, bondlist):
-    """Build the bond length array"""
-    n_bonds = np.shape(bondlist)[0]
-    xi = np.zeros(n_bonds)
-
-    for k_bond in prange(n_bonds):
-        node_i = bondlist[k_bond, 0]
-        node_j = bondlist[k_bond, 1]
-
-        xi_x = x[node_j, 0] - x[node_i, 0]
-        xi_y = x[node_j, 1] - x[node_i, 1]
-
-        xi[k_bond] = np.sqrt(xi_x**2 + xi_y**2)
-
-    return xi
-
-
-def calculate_bond_stiffness(E, delta):
-    c = (12 * E) / (np.pi * delta**4)
-    return c
 
 
 def calculate_stable_time_step(rho, dx, horizon, c):
